+++ R/blob_lease.R
@@ -28,9 +28,8 @@ acquire_lease <- function(container, blob="", duration=60, lease=NULL)
     if(!is_empty(lease))
         headers <- c(headers, list("x-ms-proposed-lease-id"=lease))
     res <- do_container_op(container, blob, options=list(comp="lease", restype="container"), headers=headers,
-                           http_verb="PUT", http_status_handler="pass")
-    httr::stop_for_status(res, storage_error_message(res))
-    httr::headers(res)[["x-ms-lease-id"]]
+                           http_verb="PUT", return_headers=TRUE)
+    res[["x-ms-lease-id"]]
 }
 
 
@@ -41,8 +40,8 @@ break_lease <- function(container, blob="", period=NULL)
     headers <- list("x-ms-lease-action"="break")
     if(!is_empty(period))
         headers <- c(headers, list("x-ms-lease-break-period"=period))
-    do_container_op(container, blob, options=list(comp="lease", restype="container"), headers=headers,
-                    http_verb="PUT")
+    invisible(do_container_op(container, blob, options=list(comp="lease", restype="container"), headers=headers,
+                              http_verb="PUT"))
 }
 
 
@@ -51,8 +50,8 @@ break_lease <- function(container, blob="", period=NULL)
 release_lease <- function(container, blob="", lease)
 {
     headers <- list("x-ms-lease-id"=lease, "x-ms-lease-action"="release")
-    do_container_op(container, blob, options=list(comp="lease", restype="container"), headers=headers,
-                    http_verb="PUT")
+    invisible(do_container_op(container, blob, options=list(comp="lease", restype="container"), headers=headers,
+                              http_verb="PUT"))
 }
 
 
@@ -61,8 +60,8 @@ release_lease <- function(container, blob="", lease)
 renew_lease <- function(container, blob="", lease)
 {
     headers <- list("x-ms-lease-id"=lease, "x-ms-lease-action"="renew")
-    do_container_op(container, blob, options=list(comp="lease", restype="container"), headers=headers,
-                    http_verb="PUT")
+    invisible(do_container_op(container, blob, options=list(comp="lease", restype="container"), headers=headers,
+                              http_verb="PUT"))
 }
 
 
@@ -72,7 +71,6 @@ change_lease <- function(container, blob="", lease, new_lease)
 {
     headers <- list("x-ms-lease-id"=lease, "x-ms-lease-action"="change", "x-ms-proposed-lease-id"=new_lease)
     res <- do_container_op(container, blob, options=list(comp="lease", restype="container"), headers=headers,
-                           http_verb="PUT", http_status_handler="pass")
-    httr::stop_for_status(res, storage_error_message(res))
-    httr::headers(res)[["x-ms-lease-id"]]
+                           http_verb="PUT", return_headers=TRUE)
+    res[["x-ms-lease-id"]]
 }